rename 'server-utils' to 'companion-client'
