+++ examples/custom-provider/client/MyCustomProvider.js
@@ -1,5 +1,5 @@
 const { Plugin } = require('@uppy/core')
-const { Provider } = require('@uppy/server-utils')
+const { Provider } = require('@uppy/companion-client')
 const ProviderViews = require('@uppy/provider-views')
 const { h } = require('preact')
