+++ packages/@uppy/companion-client/LICENSE
