+++ packages/@uppy/companion-client/README.md
@@ -1,8 +1,8 @@
-# @uppy/server-utils
+# @uppy/companion-client
 
 <img src="https://uppy.io/images/logos/uppy-dog-head-arrow.svg" width="120" alt="Uppy logo: a superman puppy in a pink suit" align="right">
 
-<a href="https://www.npmjs.com/package/@uppy/server-utils"><img src="https://img.shields.io/npm/v/@uppy/server-utils.svg?style=flat-square"></a>
+<a href="https://www.npmjs.com/package/@uppy/companion-client"><img src="https://img.shields.io/npm/v/@uppy/companion-client.svg?style=flat-square"></a>
 <a href="https://travis-ci.org/transloadit/uppy"><img src="https://img.shields.io/travis/transloadit/uppy/master.svg?style=flat-square" alt="Build Status"></a>
 
 Client library for communication with Uppy Server. Intended for use in Uppy plugins.
@@ -13,7 +13,7 @@ Uppy is being developed by the folks at [Transloadit](https://transloadit.com), 
 
 ```js
 const Uppy = require('@uppy/core')
-const { Provider, RequestClient, Socket } = require('@uppy/server-utils')
+const { Provider, RequestClient, Socket } = require('@uppy/companion-client')
 
 const uppy = Uppy()
 
@@ -35,7 +35,7 @@ socket.on('progress', () => {})
 > Unless you are writing a custom provider plugin, you do not need to install this.
 
 ```bash
-$ npm install @uppy/server-utils --save
+$ npm install @uppy/companion-client --save
 ```
 
 <!-- Undocumented currently
