+++ packages/@uppy/companion-client/src/Provider.js
