+++ packages/@uppy/companion-client/src/RequestClient.js
