+++ packages/@uppy/companion-client/src/RequestClient.test.js
