+++ packages/@uppy/companion-client/src/Socket.js
