+++ packages/@uppy/companion-client/src/Socket.test.js
