+++ packages/@uppy/companion-client/src/index.js
