+++ packages/@uppy/companion-client/types/index.d.ts
