+++ packages/@uppy/dropbox/types/index.d.ts
@@ -1,5 +1,5 @@
 import { Plugin, PluginOptions, Uppy } from '@uppy/core';
-import { ProviderOptions } from '@uppy/server-utils';
+import { ProviderOptions } from '@uppy/companion-client';
 
 export interface DropboxOptions extends PluginOptions, ProviderOptions {
   serverUrl: string;
