+++ packages/@uppy/tus/package.json
@@ -23,7 +23,7 @@
     "url": "git+https://github.com/transloadit/uppy.git"
   },
   "dependencies": {
-    "@uppy/server-utils": "0.26.0",
+    "@uppy/companion-client": "0.26.0",
     "@uppy/utils": "0.26.0",
     "tus-js-client": "^1.5.1"
   },
