+++ packages/@uppy/url/src/index.js
@@ -1,7 +1,7 @@
 const { Plugin } = require('@uppy/core')
 const Translator = require('@uppy/utils/lib/Translator')
 const { h } = require('preact')
-const { RequestClient } = require('@uppy/server-utils')
+const { RequestClient } = require('@uppy/companion-client')
 const UrlUI = require('./UrlUI.js')
 const toArray = require('@uppy/utils/lib/toArray')
