+++ packages/@uppy/xhr-upload/package.json
@@ -25,7 +25,7 @@
     "url": "git+https://github.com/transloadit/uppy.git"
   },
   "dependencies": {
-    "@uppy/server-utils": "0.26.0",
+    "@uppy/companion-client": "0.26.0",
     "@uppy/utils": "0.26.0",
     "cuid": "^2.1.1"
   },
