+++ packages/uppy/package.json
@@ -45,7 +45,7 @@
     "@uppy/progress-bar": "0.26.0",
     "@uppy/provider-views": "0.26.0",
     "@uppy/redux-dev-tools": "0.26.0",
-    "@uppy/server-utils": "0.26.0",
+    "@uppy/companion-client": "0.26.0",
     "@uppy/status-bar": "0.26.0",
     "@uppy/store-default": "0.26.0",
     "@uppy/store-redux": "0.26.0",
